+++ FtcRobotController/src/main/java/trclib/TrcDataUnwrapper.java
@@ -1,3 +1,26 @@
+/*
+ * Titan Robotics Framework Library
+ * Copyright (c) 2015 Michael H. Tsang (http://www.titanrobotics.net)
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ */
+
 package trclib;
 
 import ftclib.FtcOpMode;
@@ -195,7 +218,7 @@ public class TrcDataUnwrapper implements TrcTaskMgr.Task
                 new TrcSensor.SensorData(prevData[index].timestamp, prevData[index].value);
 
         data.value = (valueRangeHighs[index] - valueRangeLows[index])*numCrossovers[index] +
-                     (data.value - valueRangeLows[index]);
+                     ((Double)data.value - valueRangeLows[index]);
 
         if (debugEnabled)
         {
@@ -251,10 +274,10 @@ public class TrcDataUnwrapper implements TrcTaskMgr.Task
         for (int i = 0; i < numAxes; i++)
         {
             TrcSensor.SensorData data = sensor.getData(i, dataType);
-            if (Math.abs(data.value - prevData[i].value) >
+            if (Math.abs((Double)data.value - (Double)prevData[i].value) >
                 (valueRangeHighs[i] - valueRangeLows[i])/2.0)
             {
-                if (data.value > prevData[i].value)
+                if ((Double)data.value > (Double)prevData[i].value)
                 {
                     numCrossovers[i]--;
                 }
